requirements.txt
apturl==0.5.2
asgiref==3.3.4
beautifulsoup4==4.9.3
bizerror==0.4.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
crispy-tailwind==0.5.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dictop==0.1.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-flatpickr==1.0.1
django-form-admin==0.5.1
django-secure-password-input==0.1.3
django-widget-tweaks==1.4.8
dnspython==1.16.0
entrypoints==0.3
et-xmlfile==1.1.0
fastutils==0.40.2
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
jsonformatter==0.3.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
magic-import==0.1.3
Markdown==3.1.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.7
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.8.0
pyasn1==0.4.8
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.10.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pypinyin==0.41.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
xkit==0.0.0
xlsxhelper==0.3.0
